# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: EduTrack API
  version: 1.0.0-draft
  summary: Organisation task and client ticketing platform.
  description: |
    The agreed contract between all four streams. Every endpoint in blueprint §13.

    **This file is hand-authored and under review (D-001).** It is the design
    contract, not yet the generated one. From D-003 onwards `springdoc` emits the
    real spec from the Java DTOs — Bean Validation annotations are the single
    source of truth for validation rules — and D-005's CI check fails the build
    when the committed TypeScript client drifts from it. Until then this file is
    what the frontend generates against and what the backend must implement.

    ## Conventions that apply to every operation

    - **Prefix** — every path is served under `/api/v1`.
    - **Envelope** — success responses are `{ data, meta? }`. Errors are
      `{ error }` carrying an RFC 9457 problem object. Nothing returns a bare
      array or a bare scalar; adding a field later must never be a breaking change.
    - **Errors** — `application/problem+json`, RFC 9457. `type` is a stable URI
      the frontend may switch on; `detail` is for humans and may change.
    - **Idempotency** — every POST that creates something accepts
      `Idempotency-Key`. Replaying a key within 24 h returns the original response
      rather than creating a second row.
    - **ETag** — detail reads return `ETag`; send `If-None-Match` to get `304`,
...

...

...

...

...

  contact:
...
  license:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /me/password:
    patch:
      tags: [auth]
      operationId: changeOwnPassword
      summary: Change own password
      description: Rejects reuse of the last three passwords. Clears `mustChangePassword`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string }
                newPassword: { $ref: '#/components/schemas/Password' }
      responses:
        '204': { description: Changed. }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ── A-029 · two-factor enrolment ────────────────────────────────────────────
  #
  # STREAM A ADDITION — flagged for Debashis's review rather than made silently
  # (CLAUDE.md, code ownership: Stream D owns this file).
  #
  # The contract described 2FA only as LoginRequest.totpCode, with no endpoints
  # for enrolling, confirming, disabling, or for recovery codes. A-029 needs all
  # of them, and ContractConformanceTest fails the build on any served endpoint
  # that is not here — its own message says "add it to the contract and get it
  # reviewed, or remove it". These are added so the build is green and the
  # generated client can be regenerated; the review is still owed.
  #
  # Recovery codes are an addition beyond the blueprint too: §10.3 and S-04
  # describe TOTP and stop, but a second factor living on one device makes a lost
  # phone a permanently locked account.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        15-minute access token. Claims: `sub`, `role`, `permissions[]`,
        `projects[]`, `reportees[]`, `iat`, `exp`, `jti`. The refresh token is a
        separate opaque value in an HttpOnly cookie and is never sent here.
...
  responses:
    Unauthorized:
      description: Missing, expired or revoked token.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    ValidationFailed:
      description: Bean Validation failure; `errors` is field-keyed.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ValidationProblem' }
...
  schemas:
...
    Problem:
      type: object
      description: RFC 9457. `type` is stable and switchable; `detail` is for humans.
      required: [type, title, status]
      properties:
        type: { type: string, format: uri, examples: ['https://edutrack/errors/invalid-credentials'] }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        traceId: { type: string, description: Correlates with server logs and Sentry. }
    ValidationProblem:
      allOf:
        - { $ref: '#/components/schemas/Problem' }
        - type: object
          properties:
            errors:
              type: object
              additionalProperties: { type: array, items: { type: string } }
              examples: [{ title: ['must not be blank'] }]
...
    Password:
      type: string
      minLength: 8
      maxLength: 128
      description: Upper, lower, digit and symbol required; the last three are not reusable.
...